switch dynamic and static components
